requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
distro==1.9.0
groq==0.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
langdetect==1.0.9
pydantic==2.7.1
pydantic_core==2.18.2
python-telegram-bot==21.1.1
six==1.16.0
sniffio==1.3.1
typing_extensions==4.11.0
